+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Version history of koji-tool
 
-## 0.9.3 (2022-08-02)
+## 0.9.3 (2022-08-03)
 - 'builds' for a package now use specified query options
 - 'builds': add --install and --tasks options
 - 'install': update installed subpackages first then install any new subpkgs
@@ -9,6 +9,7 @@
 - 'progress' records the largest finished build.log size
 - 'progress': exclude srpm from max log sizes
 - 'progress': show pkg name instead of "unknown" nvr early on
+- 'progress': only use current time if task unfinished
 - 'tasks': fix output url selection for scratch builds
 - make parsing of states strict for earlier errors
 - error if koji userid not found